container-scripts/scripts/exports/keyvault_utils.py
import os
from azure.identity import DefaultAzureCredential
from azure.keyvault.secrets import SecretClient

def get_secret(secret_name: str) -> str:
    key_vault_name = os.environ["KEY_VAULT_NAME"]
    vault_url = f"https://{key_vault_name}.vault.azure.net"

    credential = DefaultAzureCredential()
    client = SecretClient(vault_url=vault_url, credential=credential)

    return client.get_secret(secret_name).value