container-scripts/requirements.txt
pandas
cryptography
azure-identity
azure-keyvault-secrets
azure-storage-file-datalake